requirements.txt
Django==2.1.5
django-filter==2.0.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
Markdown==3.0.1
Pillow==5.4.1
PyJWT==1.7.1
pytz==2018.9
